+++ lib/packages/wikibase/data-model-services/src/Lookup/InMemoryEntityLookup.php
@@ -5,6 +5,8 @@ namespace Wikibase\DataModel\Services\Lookup;
 use InvalidArgumentException;
 use Wikibase\DataModel\Entity\EntityDocument;
 use Wikibase\DataModel\Entity\EntityId;
+use Wikibase\DataModel\Entity\ItemId;
+use Wikibase\DataModel\Entity\PropertyId;
 
 /**
  * EntityLookup that uses an in memory array to retrieve the requested information.
@@ -18,7 +20,7 @@ use Wikibase\DataModel\Entity\EntityId;
  * @license GPL-2.0-or-later
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
  */
-class InMemoryEntityLookup implements EntityLookup {
+class InMemoryEntityLookup implements EntityLookup, ItemLookup, PropertyLookup {
 
 	/**
 	 * @var EntityDocument[]
@@ -99,6 +101,30 @@ class InMemoryEntityLookup implements EntityLookup {
 	private function throwExceptionIfNeeded( EntityId $entityId ) {
 		if ( array_key_exists( $entityId->getSerialization(), $this->exceptions ) ) {
 			throw $this->exceptions[$entityId->getSerialization()];
+		}
+	}
+
+	public function getItemForId( ItemId $itemId ) {
+		try {
+			if ( $this->hasEntity( $itemId ) ) {
+				return $this->getEntity( $itemId );
+			}
+
+			return null;
+		} catch ( \Exception $ex ) {
+			throw new ItemLookupException( $itemId, $ex->getMessage(), $ex );
+		}
+	}
+
+	public function getPropertyForId( PropertyId $propertyId ) {
+		try {
+			if ( $this->hasEntity( $propertyId ) ) {
+				return $this->getEntity( $propertyId );
+			}
+
+			return null;
+		} catch ( \Exception $ex ) {
+			throw new PropertyLookupException( $propertyId, $ex->getMessage(), $ex );
 		}
 	}